+++ app/client/src/components/editorComponents/WidgetComponentBoundary.tsx
@@ -1,7 +1,9 @@
-import { WIDGET_COMPONENT_BOUNDARY_CLASS } from "constants/componentClassNameConstants";
 import type { ReactNode } from "react";
-import React from "react";
+import React, { useContext } from "react";
 import styled from "styled-components";
+
+import WalkthroughContext from "components/featureWalkthrough/walkthroughContext";
+import { WIDGET_COMPONENT_BOUNDARY_CLASS } from "constants/componentClassNameConstants";
 
 type Props = { children: ReactNode; widgetType: string };
 
@@ -11,8 +13,24 @@ const WidgetComponentBoundaryWrapper = styled.div`
 `;
 
 function WidgetComponentBoundary(props: Props) {
+  const { isOpened: isWalkthroughOpened, popFeature } =
+    useContext(WalkthroughContext) || {};
+
+  const closeWalkthrough = () => {
+    if (isWalkthroughOpened && popFeature) {
+      popFeature("WIDGET_CONTAINER");
+    }
+  };
+
+  const onClickHandler = () => {
+    closeWalkthrough();
+  };
+
   return (
-    <WidgetComponentBoundaryWrapper className={WIDGET_COMPONENT_BOUNDARY_CLASS}>
+    <WidgetComponentBoundaryWrapper
+      className={WIDGET_COMPONENT_BOUNDARY_CLASS}
+      onClick={onClickHandler}
+    >
       {props.children}
     </WidgetComponentBoundaryWrapper>
   );
